intermediate changes for PR -rebase me

+++ docdb/docdb.go
@@ -62,7 +62,11 @@ func WithDefaultKMSKeyId(keyId string) DocDBOption {
 
 // ListDB lists documentdb clusters
 func (d *DocDB) ListDB(ctx context.Context, input *docdb.DescribeDBClustersInput) (*docdb.DescribeDBClustersOutput, error) {
-	log.Info("listing documentDB clusters")
+	if input == nil {
+		return nil, apierror.New(apierror.ErrBadRequest, "invalid input", nil)
+	}
+
+	log.Debugf("listing documentDB clusters with input %+v", input)
 
 	out, err := d.Service.DescribeDBClusters(input)
 	if err != nil {
@@ -74,6 +78,11 @@ func (d *DocDB) ListDB(ctx context.Context, input *docdb.DescribeDBClustersInput
 
 // GetDB gets information on a documentDB cluster+instance
 func (d *DocDB) GetDB(ctx context.Context, input *docdb.DescribeDBClustersInput) (*docdb.DBCluster, error) {
+	if input == nil {
+		return nil, apierror.New(apierror.ErrBadRequest, "invalid input", nil)
+	}
+
+	log.Debugf("getting documentDB cluster with input %+v", input)
 
 	out, err := d.Service.DescribeDBClusters(input)
 	if err != nil {
@@ -90,6 +99,11 @@ func (d *DocDB) GetDB(ctx context.Context, input *docdb.DescribeDBClustersInput)
 
 // CreateDBCluster creates a documentDB cluster
 func (d *DocDB) CreateDBCluster(ctx context.Context, name string, input *docdb.CreateDBClusterInput) (*docdb.DBCluster, error) {
+	if input == nil {
+		return nil, apierror.New(apierror.ErrBadRequest, "invalid input", nil)
+	}
+
+	log.Debugf("creating documentDB cluster with input %+v", input)
 
 	out, err := d.Service.CreateDBCluster(input)
 	if err != nil {
@@ -101,6 +115,11 @@ func (d *DocDB) CreateDBCluster(ctx context.Context, name string, input *docdb.C
 
 // CreateDBInstance creates a documentDB instance
 func (d *DocDB) CreateDBInstance(ctx context.Context, input *docdb.CreateDBInstanceInput) (*docdb.DBInstance, error) {
+	if input == nil {
+		return nil, apierror.New(apierror.ErrBadRequest, "invalid input", nil)
+	}
+
+	log.Debugf("creating documentDB instance with input %+v", input)
 
 	out, err := d.Service.CreateDBInstance(input)
 	if err != nil {
@@ -112,6 +131,11 @@ func (d *DocDB) CreateDBInstance(ctx context.Context, input *docdb.CreateDBInsta
 
 // DeleteDBCluster deletes a documentDB cluster
 func (d *DocDB) DeleteDBCluster(ctx context.Context, input *docdb.DeleteDBClusterInput) (*docdb.DeleteDBClusterOutput, error) {
+	if input == nil {
+		return nil, apierror.New(apierror.ErrBadRequest, "invalid input", nil)
+	}
+
+	log.Debugf("deleting documentDB cluster with input %+v", input)
 
 	out, err := d.Service.DeleteDBCluster(input)
 	if err != nil {
@@ -123,6 +147,11 @@ func (d *DocDB) DeleteDBCluster(ctx context.Context, input *docdb.DeleteDBCluste
 
 // DeleteDBInstance deletes a documentDB instance
 func (d *DocDB) DeleteDBInstance(ctx context.Context, input *docdb.DeleteDBInstanceInput) (*docdb.DeleteDBInstanceOutput, error) {
+	if input == nil {
+		return nil, apierror.New(apierror.ErrBadRequest, "invalid input", nil)
+	}
+
+	log.Debugf("deleting documentDB instance with input %+v", input)
 
 	out, err := d.Service.DeleteDBInstance(input)
 	if err != nil {
